// src/openapi/models/ArticleCommentCreationFormData.ts
// tslint:disable
/**
 * Blog
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.0
 * Contact: johan@johanbaskovec.fr
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * @export
 * @interface ArticleCommentCreationFormData
 */
export interface ArticleCommentCreationFormData {
    /**
     * @type {string}
     * @memberof ArticleCommentCreationFormData
     */
    id: string;
    /**
     * @type {string}
     * @memberof ArticleCommentCreationFormData
     */
    articleId: string;
    /**
     * @type {string}
     * @memberof ArticleCommentCreationFormData
     */
    content: string;
}